+++ CHANGELOG.md
@@ -5,6 +5,7 @@ To install or update LODA, please follow the [installation instructions](https:/
 ### Features
 
 * Support bitwise operators with negative values in formula generation
+* Support binomial with negative args in PARI code generation
 
 ### Enhancements